% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/New_BinProfet.R
\name{New_BinProfet}
\alias{New_BinProfet}
\title{Coarse Binning Variable(s)}
\usage{
New_BinProfet(
  data,
  id,
  target,
  varcol,
  min.cat = 4,
  num.bins = 10,
  min.pts.bin = 25,
  bracket = "left",
  special.values = NULL
)
}
\arguments{
\item{data}{Dataframe of that contains ID, binary target and variables to be binned.}

\item{id}{ID variable. See 'Details'.}

\item{target}{The binary target/response variable for WOE. See 'Details'.}

\item{varcol}{Vector of variables to be binned.}

\item{min.cat}{Minimum number of bins.}

\item{num.bins}{Target number of bins. Overridden by the number of levels if varcol is factor.}

\item{min.pts.bin}{Minimum number of observations in a bin.}

\item{bracket}{Indicating if the intervals should be closed on the right or left. Options include left and right.}

\item{special.values}{A vector of values that should have their own bin. See 'Details'.}
}
\value{
A dataframe with original ID and target columns and binned variable column(s).
}
\description{
Function that bins selected variable(s) and returns a dataframe with binned values. Uses greedy binning algorithm to perform coarse binning of selected variable(s).
}
\examples{
mydata <- ISLR::Default
head(mydata)
mydata$ID <- seq(1:nrow(mydata)) ## make an ID variable
mydata$default <- ifelse(mydata$default=="Yes", 1, 0) ## target coded with 1, 0
## bin balance and income
binned1 <- New_BinProfet(mydata, id="ID", target="default",
                    varcol = c("balance",  "income"), num.bins = 5)
head(binned1)
}
