% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScorecardProfet_New.R
\name{ScorecardProfet_New}
\alias{ScorecardProfet_New}
\title{Scorecard Builder}
\usage{
ScorecardProfet_New(
  object,
  id,
  target,
  varcol,
  PDO = 100,
  BaseOdds = 10,
  BasePts = 1000,
  reverse = FALSE
)
}
\arguments{
\item{object}{A WOEProfet object containing dataframes with binned and WOE values.}

\item{id}{ID variable.}

\item{target}{A binary target variable.}

\item{varcol}{Vector of WOE variables to be used in the logistic regression model.}

\item{PDO}{Points to Double Odds.}

\item{BaseOdds}{Base Odds.}

\item{BasePts}{Base Points.}

\item{reverse}{Logical. If true, higher points corresponds to a lower probability of being target.}
}
\value{
A list of a dataframe of scorecard, a scored input dataframe, and a summary of the logistic model used to built the scorecard.
}
\description{
Function that fits a logistic regression models and scores points for each bin and calculates observations' total score.
}
\examples{
#skip it for now
}
