% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreDataProfet.R
\name{ScoreDataProfet}
\alias{ScoreDataProfet}
\title{Score a Validation Data Set}
\usage{
ScoreDataProfet(data, card, id, target, varcol)
}
\arguments{
\item{data}{The validation data set, which should be binned in the same way as the scorecard in the card argument.}

\item{card}{A ScorecardProfet object. The object should be created by using the training set split from the same dataframe
as the validation set.}

\item{id}{ID variable.}

\item{target}{A binary target variable.}

\item{varcol}{Variables on the validation set to be used for scoring. WE MIGHT DETELTE THIS, NOT REALLY NEEDED}
}
\value{
A dataframe of scored validation set.
}
\description{
Function that scores the validation set using the scorecard from the ScorecardProfet object created by
the training set.
}
\examples{
#skip for now
}
