% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOE_customFac.R
\name{WOE_customFac}
\alias{WOE_customFac}
\title{Custom Binning Factor Variables}
\usage{
WOE_customFac(
  data,
  var,
  id,
  target,
  new_levels,
  color = "#0066CC",
  plot = TRUE
)
}
\arguments{
\item{data}{Dataframe containing the target variable and desired factor variables to be binned.}

\item{var}{A specific factor attribute to be binned.}

\item{id}{The unique id variable in the dataframe. Must be specified.}

\item{target}{A binary target variable. Must be specified.}

\item{new_levels}{A vector the same length as the number of levels for the categorical variable containing the new factor levels. Must be specified.}

\item{color}{A hexadecimal value representing a specific color.}

\item{plot}{Logical. The default is TRUE which generates the plots.}
}
\value{
A list contains a dataframe of the newly binned values and other information.
}
\description{
Function that bins a factor variable based on user inputted factor levels, plots the information on the new bins,
and returns a list contains a dataframe of the newly binned values and id column and more items.
}
\examples{
#ship for now
}
