% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOE_customNum.R
\name{WOE_customNum}
\alias{WOE_customNum}
\title{Custom Binning Numeric Variables}
\usage{
WOE_customNum(
  data,
  var,
  id,
  target,
  breaks,
  right_bracket = F,
  color = "#0066CC",
  plot = TRUE
)
}
\arguments{
\item{data}{Dataframe containing the target variable and desired numeric variables to be binned.}

\item{var}{A specific numeric attribute to be binned. Must be specified.}

\item{id}{The unique id variable in the dataframe. Must be specified.}

\item{target}{A binary target variable. Must be specified.}

\item{breaks}{A vector of breakpoints for the desired bins. Must be specified.}

\item{right_bracket}{Logical. Specifying whether the intervals are closed on the right or the left.}

\item{color}{A hexadecimal value representing a specific color.}

\item{plot}{Logical. The default is TRUE which generates the plots.}
}
\value{
A list contains a dataframe of the newly binned values and other information.
}
\description{
Function that bins a numeric variable based on user inputted breaks, plots the information on the new bins,
and returns a list contains a dataframe of the newly binned values and id column and more items.
}
\examples{
#skip for now
}
