% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOE_StepAIC.R
\name{WOE_StepAIC}
\alias{WOE_StepAIC}
\title{Stepwise Variable Selection}
\usage{
WOE_StepAIC(object, id, target, IVfilter, direction = "both")
}
\arguments{
\item{object}{A WOEProfet object containing dataframes with binned and WOE values.}

\item{id}{ID variable.}

\item{target}{A binary target variable.}

\item{IVfilter}{Number of desired clusters.}

\item{direction}{The mode of stepwise search direction. This should be one of "both", "backward", or "forward", with a default of "both".}
}
\value{
A list of four items.
}
\description{
Function that implements stepwise model selection by AIC to be used as a form of variable selection.
}
\examples{
##skip for now
}
