% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOEclust_hclust.R
\name{WOEclust_hclust}
\alias{WOEclust_hclust}
\title{Hierarchical Variable Clustering}
\usage{
WOEclust_hclust(object, id, target, num_clusts, method = "ward.D")
}
\arguments{
\item{object}{A WOEProfet object containing dataframes with binned and WOE values.}

\item{id}{ID variable.}

\item{target}{A binary target variable.}

\item{num_clusts}{Number of desired clusters.}

\item{method}{Clustering method to be used. This should be one of "ward.D", "ward.D2", "single", "average", "mcquitty", "median",or "centroid".}
}
\value{
A dataframe indicates the assigned clusters for the predictor variables.
}
\description{
Function that implements hierarchical clustering on the variables to be used as a form of variable selection.
}
\examples{
mydata <- ISLR::Default
mydata$ID = seq(1:nrow(mydata)) ## make the ID variable
mydata$default<-ifelse(mydata$default=="Yes",1,0) ## Creating numeric binary target variable
}
