% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOEclust_kmeans.R
\name{WOEclust_kmeans}
\alias{WOEclust_kmeans}
\title{K-means Variable Clustering}
\usage{
WOEclust_kmeans(object, id, target, num_clusts)
}
\arguments{
\item{object}{A WOEProfet object containing dataframes with binned and WOE values.}

\item{id}{ID variable.}

\item{target}{A binary target variable.}

\item{num_clusts}{Number of desired clusters.}
}
\value{
A dataframe indicates the assigned clusters for the predictor variables.
}
\description{
Function that implements kmeans variable clusteting to be used as a form of variable selection.
}
\examples{
mydata <- ISLR::Default
mydata$ID = seq(1:nrow(mydata)) ## make the ID variable
## create two new variables from bivariate normal
sigma <- matrix(c(45000,-3000,-3000, 55000), nrow = 2)
set.seed(10)
newvars <- MASS::mvrnorm(nrow(mydata),
                        mu=c(1000,200), Sigma=sigma)
}
