% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOEplotter_New.R
\name{WOEplotter_New}
\alias{WOEplotter_New}
\title{Visualizing WOE and Target Rates}
\usage{
WOEplotter_New(data, target, var, color = "#0066CC")
}
\arguments{
\item{data}{Dataframe containing binned values and a binary target variable.}

\item{target}{A numeric binary target variable.}

\item{var}{The desired WOE binned attribute to visualize.}

\item{color}{A hexadecimal value representing a specific color.}
}
\value{
three plots.
}
\description{
Function generating three plots: WOE value for each bin, target rate for each bin, and the frequency for each bin.
}
\examples{
mydata <- ISLR::Default
mydata$ID = seq(1:nrow(mydata)) ## make the ID variable
mydata$default<-ifelse(mydata$default=="Yes",1,0) ## Creating numeric binary target variable
binned <- New_BinProfet(mydata, id= "ID", target= "default", num.bins = 5) ## Binning variables
WOEplotter_New(binned, target= "default", var= "income_Bins")
}
